Refactor game-specific memory stuff to library
