+++ include/okami/devdatafinderdesc.h
@@ -1,0 +1,17 @@
+
+#include <string>
+#include <unordered_map>
+
+#include "okami/data/structs.hpp"
+
+namespace okami
+{
+extern const std::unordered_map<unsigned, std::string> animalsFedDesc;
+extern const std::unordered_map<unsigned, std::string> keyItemsFoundDesc;
+extern const std::unordered_map<unsigned, std::string> goldDustsFoundDesc;
+extern const std::unordered_map<unsigned, std::string> brushUpgradesDesc;
+extern const std::unordered_map<unsigned, std::string> commonStatesDesc;
+extern const std::unordered_map<unsigned, std::string> areasRestoredDesc;
+extern const std::unordered_map<unsigned, std::string> gameProgressDesc;
+extern const std::unordered_map<unsigned, std::string> animalsFedFirstTimeDesc;
+} // namespace okami
