+++ src/library/sources.cmake
@@ -4,6 +4,8 @@ set(library_SOURCES
   ${CMAKE_CURRENT_SOURCE_DIR}/src/library/bestiarytome.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/src/library/brushtype.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/src/library/dataverify.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/devdatafinderdesc.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/devdatamapdata.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/src/library/dojotech.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/src/library/fish.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/src/library/itemtype.cpp
